🚚 refactor: Improve orders folder structure
